refactor(capabilities): :recycle: move MotionDiv from molecules to atoms
MotionDiv is a primitive animation wrapper — it has no molecule-level
composition. Moving it to atoms/animation/motion-div removes the root
cause of atoms being forced to inline motion.div (importing a molecule
from an atom violated design-system-layering). All 8 importers updated
to the new @/components/design-system/atoms/animation/motion-div path.
